+++ RELEASE_NOTES.md
@@ -5,10 +5,13 @@
 ## General
 
 * Added support for prime/upstream kubernetes artifacts using the `config/artifacts.yaml` file
-* Added mounting for `/usr/local` for Operating System file 
+* Added mounting for `/usr/local` for Operating System file handling
+* Added ability to generate independent combustion drive
 * Add warning about providing container images with index digests
 
 ## API
+
+* Introduced `generate` command for generating combustion drives
 
 ### Image Definition Changes